+++ Demo/Application/Features/PayPal - Checkout/BraintreeDemoPayPalCheckoutViewController.swift
@@ -18,15 +18,15 @@ class BraintreeDemoPayPalCheckoutViewController: BraintreeDemoPaymentButtonBaseV
     }
     
     @objc func tappedPayPalCheckout(_ sender: UIButton) {
-        progressBlock("Tapped PayPal - Checkout using BTPayPalDriver")
+        progressBlock("Tapped PayPal - Checkout using BTPayPalClient")
         sender.setTitle("Processing...", for: .disabled)
         sender.isEnabled = false
         
-        let driver = BTPayPalDriver(apiClient: apiClient)
+        let client = BTPayPalClient(apiClient: apiClient)
         let request = BTPayPalCheckoutRequest(amount: "4.30")
         request.activeWindow = self.view.window
         
-        driver.tokenizePayPalAccount(with: request) { nonce, error in
+        client.tokenizePayPalAccount(with: request) { nonce, error in
             sender.isEnabled = true
             
             guard let nonce = nonce else {