+++ Demo/Application/Features/Preferred Payment Methods/BraintreeDemoPreferredPaymentMethodsViewController.swift
@@ -2,8 +2,8 @@ import UIKit
 
 class BraintreeDemoPreferredPaymentMethodsViewController: BraintreeDemoBaseViewController {
     private let preferredPaymentMethods: BTPreferredPaymentMethods
-    private let paypalDriver: BTPayPalDriver
-    private let venmoDriver: BTVenmoDriver
+    private let paypalClient: BTPayPalClient
+    private let venmoClient: BTVenmoClient
     private let payPalCheckoutButton = UIButton(type: .system)
     private let payPalVaultButton = UIButton(type: .system)
     private let venmoButton = UIButton(type: .system)
@@ -12,8 +12,8 @@ class BraintreeDemoPreferredPaymentMethodsViewController: BraintreeDemoBaseViewC
         guard let apiClient = BTAPIClient(authorization: authorization) else { return nil }
         
         preferredPaymentMethods = BTPreferredPaymentMethods(apiClient: apiClient)
-        paypalDriver = BTPayPalDriver(apiClient: apiClient)
-        venmoDriver = BTVenmoDriver(apiClient: apiClient)
+        paypalClient = BTPayPalClient(apiClient: apiClient)
+        venmoClient = BTVenmoClient(apiClient: apiClient)
 
         super.init(authorization: authorization)
         
@@ -75,7 +75,7 @@ class BraintreeDemoPreferredPaymentMethodsViewController: BraintreeDemoBaseViewC
         button.isEnabled = false
         
         let paypalRequest = BTPayPalCheckoutRequest(amount: "4.30")
-        paypalDriver.tokenizePayPalAccount(with: paypalRequest) { (nonce, error) in
+        paypalClient.tokenizePayPalAccount(with: paypalRequest) { (nonce, error) in
             button.isEnabled = true
 
             if let e = error {
@@ -96,7 +96,7 @@ class BraintreeDemoPreferredPaymentMethodsViewController: BraintreeDemoBaseViewC
         
         let paypalRequest = BTPayPalVaultRequest()
         paypalRequest.activeWindow = self.view.window
-        paypalDriver.tokenizePayPalAccount(with: paypalRequest) { (nonce, error) in
+        paypalClient.tokenizePayPalAccount(with: paypalRequest) { (nonce, error) in
             button.isEnabled = true
             
             if let e = error {
@@ -116,7 +116,7 @@ class BraintreeDemoPreferredPaymentMethodsViewController: BraintreeDemoBaseViewC
         button.isEnabled = false
 
         let venmoRequest = BTVenmoRequest(paymentMethodUsage: .multiUse)
-        venmoDriver.tokenizeVenmoAccount(with: venmoRequest) { (nonce, error) in
+        venmoClient.tokenizeVenmoAccount(with: venmoRequest) { (nonce, error) in
             button.isEnabled = true
             
             if let e = error {