+++ Sources/BraintreePayPal/Public/BraintreePayPal/BraintreePayPal.h
@@ -10,7 +10,7 @@ FOUNDATION_EXPORT const unsigned char BraintreePayPalVersionString[];
 #import <Braintree/BraintreeCore.h>
 #import <Braintree/BTConfiguration+PayPal.h>
 #import <Braintree/BTPayPalRequest.h>
-#import <Braintree/BTPayPalDriver.h>
+#import <Braintree/BTPayPalClient.h>
 #import <Braintree/BTPayPalAccountNonce.h>
 #import <Braintree/BTPayPalCreditFinancing.h>
 #import <Braintree/BTPayPalCreditFinancingAmount.h>
@@ -21,7 +21,7 @@ FOUNDATION_EXPORT const unsigned char BraintreePayPalVersionString[];
 #import <BraintreeCore/BraintreeCore.h>
 #import <BraintreePayPal/BTConfiguration+PayPal.h>
 #import <BraintreePayPal/BTPayPalRequest.h>
-#import <BraintreePayPal/BTPayPalDriver.h>
+#import <BraintreePayPal/BTPayPalClient.h>
 #import <BraintreePayPal/BTPayPalAccountNonce.h>
 #import <BraintreePayPal/BTPayPalCreditFinancing.h>
 #import <BraintreePayPal/BTPayPalCreditFinancingAmount.h>