+++ Sources/BraintreeThreeDSecure/BTPaymentFlowClient+ThreeDSecure.m
@@ -1,36 +1,36 @@
-#import "BTPaymentFlowDriver+ThreeDSecure_Internal.h"
+#import "BTPaymentFlowClient+ThreeDSecure_Internal.h"
 #import "BTThreeDSecureResult_Internal.h"
 #import "BTThreeDSecureRequest_Internal.h"
 #import "BTThreeDSecurePostalAddress_Internal.h"
 #import "BTThreeDSecureAdditionalInformation_Internal.h"
 
 #if __has_include(<Braintree/BraintreeThreeDSecure.h>) // CocoaPods
-#import <Braintree/BTPaymentFlowDriver+ThreeDSecure.h>
+#import <Braintree/BTPaymentFlowClient+ThreeDSecure.h>
 #import <Braintree/BTThreeDSecureRequest.h>
 #import <Braintree/BraintreeCore.h>
-#import <Braintree/BTPaymentFlowDriver_Internal.h>
+#import <Braintree/BTPaymentFlowClient_Internal.h>
 #import <Braintree/BTAPIClient_Internal.h>
 #import <Braintree/Braintree-Version.h>
 
 #elif SWIFT_PACKAGE // SPM
-#import <BraintreeThreeDSecure/BTPaymentFlowDriver+ThreeDSecure.h>
+#import <BraintreeThreeDSecure/BTPaymentFlowClient+ThreeDSecure.h>
 #import <BraintreeThreeDSecure/BTThreeDSecureRequest.h>
 #import <BraintreeCore/BraintreeCore.h>
-#import "../BraintreePaymentFlow/BTPaymentFlowDriver_Internal.h"
+#import "../BraintreePaymentFlow/BTPaymentFlowClient_Internal.h"
 #import "../BraintreeCore/BTAPIClient_Internal.h"
 #import "../BraintreeCore/Braintree-Version.h"
 
 #else // Carthage
-#import <BraintreeThreeDSecure/BTPaymentFlowDriver+ThreeDSecure.h>
+#import <BraintreeThreeDSecure/BTPaymentFlowClient+ThreeDSecure.h>
 #import <BraintreeThreeDSecure/BTThreeDSecureRequest.h>
 #import <BraintreeCore/BraintreeCore.h>
-#import <BraintreePaymentFlow/BTPaymentFlowDriver_Internal.h>
+#import <BraintreePaymentFlow/BTPaymentFlowClient_Internal.h>
 #import <BraintreeCore/BTAPIClient_Internal.h>
 #import <BraintreeCore/Braintree-Version.h>
 
 #endif
 
-@implementation BTPaymentFlowDriver (ThreeDSecure)
+@implementation BTPaymentFlowClient (ThreeDSecure)
 
 NSString * const BTThreeDSecureFlowErrorDomain = @"com.braintreepayments.BTThreeDSecureFlowErrorDomain";
 NSString * const BTThreeDSecureFlowInfoKey = @"com.braintreepayments.BTThreeDSecureFlowInfoKey";
@@ -200,11 +200,11 @@ NSString * const BTThreeDSecureFlowValidationErrorsKey = @"com.braintreepayments
     BTThreeDSecureResult *lookupResult = [[BTThreeDSecureResult alloc] initWithJSON:jsonResponse];
 
     BTThreeDSecureRequest *threeDSecureRequest = (BTThreeDSecureRequest *)request;
-    threeDSecureRequest.paymentFlowDriverDelegate = self;
+    threeDSecureRequest.paymentFlowClientDelegate = self;
 
     [self.apiClient fetchOrReturnRemoteConfiguration:^(BTConfiguration * _Nullable configuration, NSError * _Nullable configurationError) {
         if (configurationError) {
-            [threeDSecureRequest.paymentFlowDriverDelegate onPaymentComplete:nil error:configurationError];
+            [threeDSecureRequest.paymentFlowClientDelegate onPaymentComplete:nil error:configurationError];
             return;
         }
         [threeDSecureRequest processLookupResult:lookupResult configuration:configuration];