+++ Sources/BraintreeThreeDSecure/Public/BraintreeThreeDSecure/BraintreeThreeDSecure.h
@@ -12,7 +12,7 @@ FOUNDATION_EXPORT const unsigned char BraintreeThreeDSecureVersionString[];
 #import <Braintree/BraintreePaymentFlow.h>
 #import <Braintree/BraintreeCard.h>
 #import <Braintree/BTConfiguration+ThreeDSecure.h>
-#import <Braintree/BTPaymentFlowDriver+ThreeDSecure.h>
+#import <Braintree/BTPaymentFlowClient+ThreeDSecure.h>
 #import <Braintree/BTThreeDSecureAdditionalInformation.h>
 #import <Braintree/BTThreeDSecureLookup.h>
 #import <Braintree/BTThreeDSecurePostalAddress.h>
@@ -29,7 +29,7 @@ FOUNDATION_EXPORT const unsigned char BraintreeThreeDSecureVersionString[];
 #import <BraintreePaymentFlow/BraintreePaymentFlow.h>
 #import <BraintreeCard/BraintreeCard.h>
 #import <BraintreeThreeDSecure/BTConfiguration+ThreeDSecure.h>
-#import <BraintreeThreeDSecure/BTPaymentFlowDriver+ThreeDSecure.h>
+#import <BraintreeThreeDSecure/BTPaymentFlowClient+ThreeDSecure.h>
 #import <BraintreeThreeDSecure/BTThreeDSecureAdditionalInformation.h>
 #import <BraintreeThreeDSecure/BTThreeDSecureLookup.h>
 #import <BraintreeThreeDSecure/BTThreeDSecurePostalAddress.h>