+++ UnitTests/BraintreePayPalTests/BraintreePayPalTests-Bridging-Header.h
@@ -1,4 +1,4 @@
-#import <BraintreePayPal/BTPayPalDriver_Internal.h>
+#import <BraintreePayPal/BTPayPalClient_Internal.h>
 #import <BraintreePayPal/BTPayPalRequest_Internal.h>
 #import <BraintreePayPal/BTPayPalCheckoutRequest_Internal.h>
 #import <BraintreeCore/BTAPIClient_Internal.h>