+++ UnitTests/BraintreeTestShared/MockViewControllerPresentingDelegate.swift
@@ -5,16 +5,16 @@ import BraintreeCore
     public var requestsPresentationOfViewControllerExpectation : XCTestExpectation? = nil
     public var requestsDismissalOfViewControllerExpectation : XCTestExpectation? = nil
     public var lastViewController : UIViewController? = nil
-    public var lastPaymentDriver : AnyObject? = nil
+    public var lastPaymentClient : AnyObject? = nil
 
-    public func paymentDriver(_ driver: Any, requestsDismissalOf viewController: UIViewController) {
-        lastPaymentDriver = driver as AnyObject?
+    public func paymentClient(_ client: Any, requestsDismissalOf viewController: UIViewController) {
+        lastPaymentClient = client as AnyObject?
         lastViewController = viewController
         requestsDismissalOfViewControllerExpectation?.fulfill()
     }
 
-    public func paymentDriver(_ driver: Any, requestsPresentationOf viewController: UIViewController) {
-        lastPaymentDriver = driver as AnyObject?
+    public func paymentClient(_ client: Any, requestsPresentationOf viewController: UIViewController) {
+        lastPaymentClient = client as AnyObject?
         lastViewController = viewController
         requestsPresentationOfViewControllerExpectation?.fulfill()
     }