+++ UnitTests/BraintreeThreeDSecureTests/BTThreeDSecure_Tests.swift
@@ -31,12 +31,12 @@ class BTThreeDSecure_UnitTests: XCTestCase {
         mockAPIClient.cannedConfigurationResponseBody = BTJSON(value: [
             "assetsUrl": "http://assets.example.com"
         ])
-        let driver = BTPaymentFlowDriver(apiClient: mockAPIClient)
-        driver.viewControllerPresentingDelegate = viewControllerPresentingDelegate
+        let client = BTPaymentFlowClient(apiClient: mockAPIClient)
+        client.viewControllerPresentingDelegate = viewControllerPresentingDelegate
 
         mockAPIClient.cannedResponseBody = BTJSON(value: getAuthRequiredLookupResponse())
 
-        driver.startPaymentFlow(threeDSecureRequest) { (result, error) in
+        client.startPaymentFlow(threeDSecureRequest) { (result, error) in
             
         }
         
@@ -53,9 +53,9 @@ class BTThreeDSecure_UnitTests: XCTestCase {
             "threeDSecure": [],
             "assetsUrl": "http://assets.example.com"
             ])
-        let driver = BTPaymentFlowDriver(apiClient: mockAPIClient)
+        let client = BTPaymentFlowClient(apiClient: mockAPIClient)
 
-        driver.startPaymentFlow(threeDSecureRequest) { (result, error) in
+        client.startPaymentFlow(threeDSecureRequest) { (result, error) in
             XCTAssertNotNil(error)
             XCTAssertNil(result)
             guard let error = error as NSError? else {return}
@@ -79,9 +79,9 @@ class BTThreeDSecure_UnitTests: XCTestCase {
             "threeDSecure": ["cardinalAuthenticationJWT": "FAKE_JWT"],
             "assetsUrl": "http://assets.example.com"
             ])
-        let driver = BTPaymentFlowDriver(apiClient: mockAPIClient)
+        let client = BTPaymentFlowClient(apiClient: mockAPIClient)
 
-        driver.startPaymentFlow(threeDSecureRequest) { (result, error) in
+        client.startPaymentFlow(threeDSecureRequest) { (result, error) in
             XCTAssertNotNil(error)
             XCTAssertNil(result)
             guard let error = error as NSError? else { return }
@@ -106,9 +106,9 @@ class BTThreeDSecure_UnitTests: XCTestCase {
             "threeDSecure": ["cardinalAuthenticationJWT": "FAKE_JWT"],
             "assetsUrl": "http://assets.example.com"
             ])
-        let driver = BTPaymentFlowDriver(apiClient: mockAPIClient)
+        let client = BTPaymentFlowClient(apiClient: mockAPIClient)
 
-        driver.startPaymentFlow(threeDSecureRequest) { (result, error) in
+        client.startPaymentFlow(threeDSecureRequest) { (result, error) in
             XCTAssertNotNil(error)
             XCTAssertNil(result)
             guard let error = error as NSError? else { return }
@@ -132,8 +132,8 @@ class BTThreeDSecure_UnitTests: XCTestCase {
             "threeDSecure": ["cardinalAuthenticationJWT": "FAKE_JWT"],
             "assetsUrl": "http://assets.example.com"
             ])
-        let driver = BTPaymentFlowDriver(apiClient: mockAPIClient)
-        driver.viewControllerPresentingDelegate = viewControllerPresentingDelegate
+        let client = BTPaymentFlowClient(apiClient: mockAPIClient)
+        client.viewControllerPresentingDelegate = viewControllerPresentingDelegate
         let responseBody = [
             "paymentMethod": [
                 "consumed": false,
@@ -159,7 +159,7 @@ class BTThreeDSecure_UnitTests: XCTestCase {
             ] as [String : Any]
         mockAPIClient.cannedResponseBody = BTJSON(value: responseBody)
 
-        driver.startPaymentFlow(threeDSecureRequest) { (result, error) in
+        client.startPaymentFlow(threeDSecureRequest) { (result, error) in
             guard let result = result as? BTThreeDSecureResult else { XCTFail(); return }
             guard let tokenizedCard = result.tokenizedCard else { XCTFail(); return }
 
@@ -187,10 +187,10 @@ class BTThreeDSecure_UnitTests: XCTestCase {
             "threeDSecure": ["cardinalAuthenticationJWT": "FAKE_JWT"],
             "assetsUrl": "http://assets.example.com"
             ])
-        let driver = BTPaymentFlowDriver(apiClient: mockAPIClient)
-        driver.viewControllerPresentingDelegate = viewControllerPresentingDelegate
+        let client = BTPaymentFlowClient(apiClient: mockAPIClient)
+        client.viewControllerPresentingDelegate = viewControllerPresentingDelegate
 
-        driver.startPaymentFlow(threeDSecureRequest) { (result, error) in
+        client.startPaymentFlow(threeDSecureRequest) { (result, error) in
             expectation.fulfill()
         }
 
@@ -207,10 +207,10 @@ class BTThreeDSecure_UnitTests: XCTestCase {
             "threeDSecure": ["cardinalAuthenticationJWT": "FAKE_JWT"],
             "assetsUrl": "http://assets.example.com"
             ])
-        let driver = BTPaymentFlowDriver(apiClient: mockAPIClient)
-        driver.viewControllerPresentingDelegate = viewControllerPresentingDelegate
+        let client = BTPaymentFlowClient(apiClient: mockAPIClient)
+        client.viewControllerPresentingDelegate = viewControllerPresentingDelegate
 
-        driver.startPaymentFlow(threeDSecureRequest) { (result, error) in
+        client.startPaymentFlow(threeDSecureRequest) { (result, error) in
             XCTAssertNotNil(error)
             XCTAssertNil(result)
             guard let error = error as NSError? else {return}
@@ -231,13 +231,13 @@ class BTThreeDSecure_UnitTests: XCTestCase {
         let viewControllerPresentingDelegate = MockViewControllerPresentingDelegate()
         viewControllerPresentingDelegate.requestsPresentationOfViewControllerExpectation = self.expectation(description: "Delegate received requestsPresentationOfViewController")
         
-        let driver = BTPaymentFlowDriver(apiClient: mockAPIClient)
-        driver.viewControllerPresentingDelegate = viewControllerPresentingDelegate
+        let client = BTPaymentFlowClient(apiClient: mockAPIClient)
+        client.viewControllerPresentingDelegate = viewControllerPresentingDelegate
 
         mockAPIClient.cannedResponseBody = BTJSON(value: getAuthRequiredLookupResponse())
         
         var paymentFinishedExpectation: XCTestExpectation? = nil
-        driver.startPaymentFlow(threeDSecureRequest) { (result, error) in
+        client.startPaymentFlow(threeDSecureRequest) { (result, error) in
             XCTAssertNil(error)
             XCTAssertNotNil(result)
             paymentFinishedExpectation!.fulfill()
@@ -246,7 +246,7 @@ class BTThreeDSecure_UnitTests: XCTestCase {
         waitForExpectations(timeout: 2, handler: nil)
         
         paymentFinishedExpectation = self.expectation(description: "Start payment expectation")
-        BTPaymentFlowDriver.handleReturnURL(URL(string: "com.braintreepayments.demo.payments://x-callback-url/braintree/threedsecure?auth_response=%7B%22paymentMethod%22:%7B%22type%22:%22CreditCard%22,%22nonce%22:%220d3e1cc8-50a4-0437-720b-c03c722f0d0a%22,%22description%22:%22ending+in+02%22,%22consumed%22:false,%22threeDSecureInfo%22:%7B%22liabilityShifted%22:true,%22liabilityShiftPossible%22:true,%22status%22:%22authenticate_successful%22,%22enrolled%22:%22Y%22%7D,%22details%22:%7B%22lastTwo%22:%2202%22,%22lastFour%22:%220002%22,%22cardType%22:%22Visa%22%7D,%22bin_data%22:%7B%22prepaid%22:%22Unknown%22,%22healthcare%22:%22Unknown%22,%22debit%22:%22Unknown%22,%22durbin_regulated%22:%22Unknown%22,%22commercial%22:%22Unknown%22,%22payroll%22:%22Unknown%22,%22issuing_bank%22:%22Unknown%22,%22country_of_issuance%22:%22Unknown%22,%22product_id%22:%22Unknown%22%7D%7D,%22threeDSecureInfo%22:%7B%22liabilityShifted%22:true,%22liabilityShiftPossible%22:true%7D,%22success%22:true%7D")!)
+        BTPaymentFlowClient.handleReturnURL(URL(string: "com.braintreepayments.demo.payments://x-callback-url/braintree/threedsecure?auth_response=%7B%22paymentMethod%22:%7B%22type%22:%22CreditCard%22,%22nonce%22:%220d3e1cc8-50a4-0437-720b-c03c722f0d0a%22,%22description%22:%22ending+in+02%22,%22consumed%22:false,%22threeDSecureInfo%22:%7B%22liabilityShifted%22:true,%22liabilityShiftPossible%22:true,%22status%22:%22authenticate_successful%22,%22enrolled%22:%22Y%22%7D,%22details%22:%7B%22lastTwo%22:%2202%22,%22lastFour%22:%220002%22,%22cardType%22:%22Visa%22%7D,%22bin_data%22:%7B%22prepaid%22:%22Unknown%22,%22healthcare%22:%22Unknown%22,%22debit%22:%22Unknown%22,%22durbin_regulated%22:%22Unknown%22,%22commercial%22:%22Unknown%22,%22payroll%22:%22Unknown%22,%22issuing_bank%22:%22Unknown%22,%22country_of_issuance%22:%22Unknown%22,%22product_id%22:%22Unknown%22%7D%7D,%22threeDSecureInfo%22:%7B%22liabilityShifted%22:true,%22liabilityShiftPossible%22:true%7D,%22success%22:true%7D")!)
         
         waitForExpectations(timeout: 2, handler: nil)
     }
@@ -259,13 +259,13 @@ class BTThreeDSecure_UnitTests: XCTestCase {
         let viewControllerPresentingDelegate = MockViewControllerPresentingDelegate()
         viewControllerPresentingDelegate.requestsPresentationOfViewControllerExpectation = self.expectation(description: "Delegate received requestsPresentationOfViewController")
         
-        let driver = BTPaymentFlowDriver(apiClient: mockAPIClient)
-        driver.viewControllerPresentingDelegate = viewControllerPresentingDelegate
+        let client = BTPaymentFlowClient(apiClient: mockAPIClient)
+        client.viewControllerPresentingDelegate = viewControllerPresentingDelegate
         
         mockAPIClient.cannedResponseBody = BTJSON(value: getAuthRequiredLookupResponse())
         
         var paymentFinishedExpectation: XCTestExpectation? = nil
-        driver.startPaymentFlow(threeDSecureRequest) { (result, error) in
+        client.startPaymentFlow(threeDSecureRequest) { (result, error) in
             XCTAssertNotNil(error)
             XCTAssertNil(result)
             guard let error = error as NSError? else {return}
@@ -277,7 +277,7 @@ class BTThreeDSecure_UnitTests: XCTestCase {
         waitForExpectations(timeout: 2, handler: nil)
         
         paymentFinishedExpectation = self.expectation(description: "Start payment expectation")
-        BTPaymentFlowDriver.handleReturnURL(URL(string: "com.braintreepayments.demo.payments://x-callback-url/braintree/threedsecure?auth_response=%7B%22threeDSecureInfo%22:%7B%22liabilityShifted%22:false,%22liabilityShiftPossible%22:true%7D,%22error%22:%7B%22message%22:%22Failed+to+authenticate,+please+try+a+different+form+of+payment.%22%7D,%22success%22:false%7D")!)
+        BTPaymentFlowClient.handleReturnURL(URL(string: "com.braintreepayments.demo.payments://x-callback-url/braintree/threedsecure?auth_response=%7B%22threeDSecureInfo%22:%7B%22liabilityShifted%22:false,%22liabilityShiftPossible%22:true%7D,%22error%22:%7B%22message%22:%22Failed+to+authenticate,+please+try+a+different+form+of+payment.%22%7D,%22success%22:false%7D")!)
         
         waitForExpectations(timeout: 2, handler: nil)
     }
@@ -290,13 +290,13 @@ class BTThreeDSecure_UnitTests: XCTestCase {
         let viewControllerPresentingDelegate = MockViewControllerPresentingDelegate()
         viewControllerPresentingDelegate.requestsPresentationOfViewControllerExpectation = self.expectation(description: "Delegate received requestsPresentationOfViewController")
 
-        let driver = BTPaymentFlowDriver(apiClient: mockAPIClient)
-        driver.viewControllerPresentingDelegate = viewControllerPresentingDelegate
+        let client = BTPaymentFlowClient(apiClient: mockAPIClient)
+        client.viewControllerPresentingDelegate = viewControllerPresentingDelegate
 
         mockAPIClient.cannedResponseBody = BTJSON(value: getAuthRequiredLookupResponse())
 
         var paymentFinishedExpectation: XCTestExpectation? = nil
-        driver.startPaymentFlow(threeDSecureRequest) { (result, error) in
+        client.startPaymentFlow(threeDSecureRequest) { (result, error) in
             XCTAssertNotNil(error)
             XCTAssertNil(result)
             guard let error = error as NSError? else {return}
@@ -309,7 +309,7 @@ class BTThreeDSecure_UnitTests: XCTestCase {
         waitForExpectations(timeout: 2, handler: nil)
 
         paymentFinishedExpectation = self.expectation(description: "Start payment expectation")
-        BTPaymentFlowDriver.handleReturnURL(URL(string: "com.braintreepayments.demo.payments://x-callback-url/braintree/threedsecure?no-auth=bad-response")!)
+        BTPaymentFlowClient.handleReturnURL(URL(string: "com.braintreepayments.demo.payments://x-callback-url/braintree/threedsecure?no-auth=bad-response")!)
 
         waitForExpectations(timeout: 2, handler: nil)
 
@@ -324,13 +324,13 @@ class BTThreeDSecure_UnitTests: XCTestCase {
         let viewControllerPresentingDelegate = MockViewControllerPresentingDelegate()
         viewControllerPresentingDelegate.requestsPresentationOfViewControllerExpectation = self.expectation(description: "Delegate received requestsPresentationOfViewController")
 
-        let driver = BTPaymentFlowDriver(apiClient: mockAPIClient)
-        driver.viewControllerPresentingDelegate = viewControllerPresentingDelegate
+        let client = BTPaymentFlowClient(apiClient: mockAPIClient)
+        client.viewControllerPresentingDelegate = viewControllerPresentingDelegate
 
         mockAPIClient.cannedResponseBody = BTJSON(value: getAuthRequiredLookupResponse())
 
         var paymentFinishedExpectation: XCTestExpectation? = nil
-        driver.startPaymentFlow(threeDSecureRequest) { (result, error) in
+        client.startPaymentFlow(threeDSecureRequest) { (result, error) in
             XCTAssertNotNil(error)
             XCTAssertNil(result)
             guard let error = error as NSError? else {return}
@@ -343,7 +343,7 @@ class BTThreeDSecure_UnitTests: XCTestCase {
         waitForExpectations(timeout: 2, handler: nil)
 
         paymentFinishedExpectation = self.expectation(description: "Start payment expectation")
-        BTPaymentFlowDriver.handleReturnURL(URL(string: "com.braintreepayments.demo.payments://x-callback-url/braintree/threedsecure?auth_response=%7B%22paymentMethod%22:%7B%22type%22:%22CreditCard%22,%22nonce%22:%220d3e1cc8-50a4-0437-720b-c03c722f0d0a%22,BAD-JSON")!)
+        BTPaymentFlowClient.handleReturnURL(URL(string: "com.braintreepayments.demo.payments://x-callback-url/braintree/threedsecure?auth_response=%7B%22paymentMethod%22:%7B%22type%22:%22CreditCard%22,%22nonce%22:%220d3e1cc8-50a4-0437-720b-c03c722f0d0a%22,BAD-JSON")!)
 
         waitForExpectations(timeout: 2, handler: nil)
 
@@ -358,13 +358,13 @@ class BTThreeDSecure_UnitTests: XCTestCase {
         let viewControllerPresentingDelegate = MockViewControllerPresentingDelegate()
         viewControllerPresentingDelegate.requestsPresentationOfViewControllerExpectation = self.expectation(description: "Delegate received requestsPresentationOfViewController")
 
-        let driver = BTPaymentFlowDriver(apiClient: mockAPIClient)
-        driver.viewControllerPresentingDelegate = viewControllerPresentingDelegate
+        let client = BTPaymentFlowClient(apiClient: mockAPIClient)
+        client.viewControllerPresentingDelegate = viewControllerPresentingDelegate
 
         mockAPIClient.cannedResponseBody = BTJSON(value: getAuthRequiredLookupResponse())
 
         var paymentFinishedExpectation: XCTestExpectation? = nil
-        driver.startPaymentFlow(threeDSecureRequest) { (result, error) in
+        client.startPaymentFlow(threeDSecureRequest) { (result, error) in
             XCTAssertNotNil(error)
             XCTAssertNil(result)
             guard let error = error as NSError? else {return}
@@ -377,7 +377,7 @@ class BTThreeDSecure_UnitTests: XCTestCase {
         waitForExpectations(timeout: 2, handler: nil)
 
         paymentFinishedExpectation = self.expectation(description: "Start payment expectation")
-        BTPaymentFlowDriver.handleReturnURL(URL(string: "com.braintreepayments.demo.payments://x-callback-url/braintree/threedsecure?auth_response=%22STRING%22")!)
+        BTPaymentFlowClient.handleReturnURL(URL(string: "com.braintreepayments.demo.payments://x-callback-url/braintree/threedsecure?auth_response=%22STRING%22")!)
 
         waitForExpectations(timeout: 2, handler: nil)
 
@@ -426,8 +426,8 @@ class BTThreeDSecure_UnitTests: XCTestCase {
         mockAPIClient.cannedConfigurationResponseBody = BTJSON(value: [
             "assetsUrl": "http://assets.example.com",
             ])
-        let driver = BTPaymentFlowDriver(apiClient: mockAPIClient)
-        driver.viewControllerPresentingDelegate = viewControllerPresentingDelegate
+        let client = BTPaymentFlowClient(apiClient: mockAPIClient)
+        client.viewControllerPresentingDelegate = viewControllerPresentingDelegate
         let responseBody = [
             "paymentMethod": [
                 "consumed": false,
@@ -460,7 +460,7 @@ class BTThreeDSecure_UnitTests: XCTestCase {
             ] as [String : Any]
         mockAPIClient.cannedResponseBody = BTJSON(value: responseBody)
 
-        driver.startPaymentFlow(threeDSecureRequest) { (result, error) in
+        client.startPaymentFlow(threeDSecureRequest) { (result, error) in
 
         }
 
@@ -477,11 +477,11 @@ class BTThreeDSecure_UnitTests: XCTestCase {
         mockAPIClient.cannedConfigurationResponseBody = BTJSON(value: [
             "assetsUrl": "http://assets.example.com",
             ])
-        let driver = BTPaymentFlowDriver(apiClient: mockAPIClient)
+        let client = BTPaymentFlowClient(apiClient: mockAPIClient)
         mockAPIClient.cannedResponseError = NSError(domain:"BTError", code: 500, userInfo: nil)
 
         let expectation = self.expectation(description: "Start payment expectation")
-        driver.startPaymentFlow(threeDSecureRequest) { (result, error) in
+        client.startPaymentFlow(threeDSecureRequest) { (result, error) in
             guard (error as NSError?) != nil else {return}
             expectation.fulfill()
         }
@@ -501,14 +501,14 @@ class BTThreeDSecure_UnitTests: XCTestCase {
             "threeDSecure": ["cardinalAuthenticationJWT": "FAKE_JWT"],
             "assetsUrl": "http://assets.example.com"
             ])
-        let driver = BTPaymentFlowDriver(apiClient: mockAPIClient)
+        let client = BTPaymentFlowClient(apiClient: mockAPIClient)
 
         let expectation = self.expectation(description: "willCallCompletion")
 
         threeDSecureRequest.nonce = "fake-card-nonce"
         threeDSecureRequest.dfReferenceID = "fake-df-reference-id"
 
-        driver.prepareLookup(threeDSecureRequest) { (clientData, error) in
+        client.prepareLookup(threeDSecureRequest) { (clientData, error) in
             XCTAssertNil(error)
             XCTAssertNotNil(clientData)
             if let data = clientData!.data(using: .utf8) {