+++ UnitTests/BraintreeVenmoTests/BraintreeVenmoTests-Bridging-Header.h
@@ -1,4 +1,4 @@
-#import <BraintreeVenmo/BTVenmoDriver_Internal.h>
+#import <BraintreeVenmo/BTVenmoClient_Internal.h>
 #import <BraintreeVenmo/BTVenmoAppSwitchReturnURL.h>
 #import <BraintreeVenmo/BTVenmoAppSwitchRequestURL.h>
 #import <BraintreeVenmo/BTVenmoAccountNonce_Internal.h>